set default page to recording module
